Split up checkbox tests into their own files
