+++ apps/fluent-tester/src/TestComponents/CheckboxV1/SizeCheckboxTest.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+import { Checkbox } from '@fluentui-react-native/experimental-checkbox';
+
+export const SizeCheckbox: React.FunctionComponent = () => {
+  return (
+    <>
+      <Checkbox tooltip="Medium checkbox" size="medium" />
+      <Checkbox tooltip="Large checkbox" size="large" />
+      <Checkbox label="Medium checkbox" size="medium" />
+      <Checkbox label="Large checkbox" size="large" />
+    </>
+  );
+};
